requirements/requirements-llm_model.txt
langchain-ollama
llama-index-llms-langchain
